requirements.txt
biopython>=1.83
pandas>=2.2
numpy>=1.26
pytest>=8.0
